// src/types/playbook_run.ts
export interface PlaybookRun {
  id: string;
  name: string;
  summary: string;
  summary_modified_at: number;
  owner_user_id: string;
  reporter_user_id: string;
  team_id: string;
  channel_id: string;
  create_at: number;
  end_at: number;
  post_id: string;
  playbook_id: string;
  checklists: Checklist[];
  status_posts: StatusPost[];
  current_status: PlaybookRunStatus;
  last_status_update_at: number;
  reminder_post_id: string;
  reminder_message_template: string;
  reminder_timer_default_seconds: number;
  status_update_enabled: boolean;
  broadcast_channel_ids: string[];
  status_update_broadcast_webhooks_enabled: boolean;
  webhook_on_status_update_urls: string[];

  /** Whether run updates should be broadcasted to channels */
  status_update_broadcast_channels_enabled: boolean;

  /** Previous reminder timer as nanoseconds */
  previous_reminder: number;
  timeline_events: TimelineEvent[];
  retrospective: string;
  retrospective_published_at: number;
  retrospective_was_canceled: boolean;
  retrospective_reminder_interval_seconds: number;
  retrospective_enabled: boolean;
  participant_ids: string[];
  metrics_data: RunMetricData[];

  /** Whether a channel member should be created when a new participant joins the run */
  create_channel_member_on_new_participant: boolean;

  /** Whether a channel member should be removed when an existing participant leaves the run */
  remove_channel_member_on_removed_participant: boolean;

  // type: PlaybookRunType
}

export interface LimitedUser {
  user_id: string;
  first_name: string;
  last_name: string;
}

export interface LimitedPost {
  message: string;
  create_at: number;
  user_id: string;
}

export interface StatusPost {
  id: string;
  create_at: number;
  delete_at: number;
}
// ...
export enum PlaybookRunStatus {
  InProgress = 'InProgress',
  Finished = 'Finished',
}

export interface RunMetricData {
  metric_config_id: string;
  value: number | null;
}